Issue #3244533 by andypost, daffie: usleep() usage triggers deprecations on PHP 8.1

+++ core/tests/Drupal/KernelTests/Core/File/DirectoryTest.php
@@ -224,7 +224,7 @@ class DirectoryTest extends FileTestBase {
       elseif ($pid == 0) {
         // Sleep so that all the forks start preparing the directory at the same
         // time.
-        usleep(($time_to_start - microtime(TRUE)) * 1000000);
+        usleep((int) (($time_to_start - microtime(TRUE)) * 1000000));
         $file_system->prepareDirectory($directory, FileSystemInterface::CREATE_DIRECTORY);
         exit();
       }
